Structural conflicts: main's hfresh temp-vector/bucket-view refactor
(TempVectorForIDWithViewThunk, fetchNormalizedVector, parallel rescore
with pooled buffers) landed via the stable-branch merge chain and
rewrote the code this PR's changes live in. Resolution ports this PR's
semantics onto the refactored fetch path rather than picking sides:

- config/shard wiring: keep both thunks — MultiVectorForIDThunk (ours)
  and TempVectorForIDWithViewThunk (main's), plus main's nil-check.
- HFresh struct: vectorForID/multiVectorForID (ours) coexist with
  tempVectorForIDThunk/tempVectors (main's).
- SearchByVector: muvera early-return moved back BEFORE main's parallel
  rescore machinery (auto-merge placed it after, which would have done
  wasted single-vector fetches for multi-vector docs); max(k,
  rescoreLimit) budget, atomic knob loads, selectCentroids extraction
  (ours) merged with slow-query annotations and the parallel
  bucket-view rescore (main's). The stale-entry ErrNotFound skip lives
  on in main's worker loop.
- distToNode: main's with-view signature wins (our change there was a
  cosmetic rename superseded by the refactor).
- normalizeVecInPlace: use the precomputed needsNormalization flag
  (same condition, one canonical check).
- test fixtures: newTestIndex gains setDelegatingTempThunk +
  createObjectsBucket; populateRandomSingleVectorIndex assigns the
  vector thunk via cfg so the delegating temp thunk sees it.

+++ adapters/repos/db/vector/hnsw/distancer/asm/l2_arm64.go
@@ -156,29 +156,30 @@ func L2ByteARM64(x []uint8, y []uint8) uint32 {
 	case 4:
 		return l24[uint8, uint32](x, y)
 	case 6:
-		// manually inlined l26(x, y)
-		diff := x[5] - y[5]
+		// manually inlined l26(x, y). The differences must be widened before
+		// squaring: uint8 arithmetic would truncate the squares mod 256.
+		diff := uint32(x[5]) - uint32(y[5])
 		sum := diff * diff
 
-		diff = x[4] - y[4]
+		diff = uint32(x[4]) - uint32(y[4])
 		sum += diff * diff
 
-		return l24[uint8, uint32](x, y) + uint32(sum)
+		return l24[uint8, uint32](x, y) + sum
 	case 8:
 		// manually inlined l28(x, y)
-		diff := x[7] - y[7]
+		diff := uint32(x[7]) - uint32(y[7])
 		sum := diff * diff
 
-		diff = x[6] - y[6]
+		diff = uint32(x[6]) - uint32(y[6])
 		sum += diff * diff
 
-		diff = x[5] - y[5]
+		diff = uint32(x[5]) - uint32(y[5])
 		sum += diff * diff
 
-		diff = x[4] - y[4]
+		diff = uint32(x[4]) - uint32(y[4])
 		sum += diff * diff
 
-		return l24[uint8, uint32](x, y) + uint32(sum)
+		return l24[uint8, uint32](x, y) + sum
 	case 10:
 		return l210[uint8, uint32](x, y)
 	case 12:
